QuickActions refactoring
